--- a/src/api/GitHubAPI.test.ts
+++ b/src/api/GitHubAPI.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi } from "vitest";
+import { describe, expect, it, vi } from "vitest";
 
 import GitHubAPI from "./GitHubAPI";
 import { ProjectsJSON } from "../projects/list/ProjectList";