… .github/dependabot.yml
… package.json
… src/projects/Projects.test.tsx