## Problem

The GitHub API is called during every project test, causing the quota to be quickly reached. This floods the logs with extraneous error messages and slows down the test runner.

## Solution

I mocked the `Octokit` methods by putting the API logic into a separate class. This cleans up the `ProjectList` component and makes it easier to mock the API calls since the logic isn't solely contained in a hook.

## Test Plan

All the tests pass without any errors or warnings. The projects tab continues to work as expected. I'll compare the performance of the tests from before to see if there's any improvement. (about half a second improvement)

| Test | Before | After | Difference |
| - | - | - | - |
| npm test | 14s | 14s | 0s |
| ProjectList.test.tsx | 1720ms | 2061ms | +341ms |
| Projects.test.tsx | 2318ms | 1810ms | -508ms |
| ProjectDetails.test.tsx | 11779ms | 11406ms | -373ms |
| GitHubAPI.test.ts | N/A | 18ms | +18ms |